Stage 4 rework: V2-049, V2-050, V2-051
V2-051 -- the route resolved the team by global primary key, so one game's URL
reached another game's team. Neither guard supplies that relationship: the
student guard proves membership in the URL team, and the instructor guard
passes because the actor really does own the URL game. _get_team() now resolves
through the game (all nine call sites), and the product and platform resolve
through the team. A cross-cohort attempt on this mutating route is recorded as
an authorization refusal, which nothing was doing before -- the response stays
a uniform 404 so the record costs the caller no information.

V2-050 -- brand awareness filtered historical marketing rows on the product's
live platform foreign key while comparing it against a round-resolved platform.
After a switch those can never agree, so every qualifying row dropped out and a
past round's awareness fell to zero. Nothing raised: an emptied sum is still a
number. Each row is now attributed to the platform its product used in that
row's own round.

The Stage 4 inventory missed this because it looked for direct attribute reads
and resolved_platform() calls. A static guard now covers relationship
traversals too, and it immediately found a second latent copy of the same shape
in _team_has_generation() -- uncalled, so not a live defect, but corrected
rather than left as the seed of a repeat.

V2-049 -- the write-off reached context.opex and stopped there. The statement
assembly enumerated six other opex keys by name, so the team was charged
nothing, shown nothing, and taxed as if the switch had not happened. It now
joins total_opex (reaching operating income, net income and cash), the tax
deductions, and RoundResultFinancials as its own field, exposed on the team's
own results but deliberately not on the competitor block. Decimal units are
preserved end to end; the integer cast was losing 100.50 units as 100.

The new competitive field and Stage 4's earlier history section both change
what a round hashes to, so MANIFEST_SCHEMA_VERSION moves to 3 and the reviewed
inventory is written to manifest_schema_v3.json, keeping v2's file as the
record of what v2 meant. Two envelope definitions sharing one version is the
one thing require_schema_version cannot survive: it would compare a stored hash
against a differently-defined envelope of the same version and read the
difference as tampering.

Also commits the auditor's rework report and register entries.

+++ backend/core/models/team_state.py
@@ -383,7 +383,8 @@ class TeamProductPlatformHistory(models.Model):
     switched_at = models.DateTimeField(auto_now_add=True)
     #: What the switch cost the team, so the write-off stays reconstructable
     #: beside the association that caused it.
-    inventory_written_off_units = models.IntegerField(default=0)
+    inventory_written_off_units = models.DecimalField(
+        max_digits=15, decimal_places=2, default=0)
     inventory_write_off = models.DecimalField(
         max_digits=15, decimal_places=2, default=0)
